Use toError
